+++ react/components/HighStock.tsx
@@ -1,17 +1,17 @@
 import React from 'react'
-import type { FunctionComponent } from 'react'
-import classNames from 'classnames'
 
-import styles from '../styles.css'
+import { useContextCssHandles } from './CssHandlesContext'
 
 interface Props {
   text: string
 }
 
-const HighStock: FunctionComponent<Props> = ({ text }) => (
-  <span className={classNames(styles.highStockText, 'c-muted-2 t-body')}>
-    {text}
-  </span>
-)
+function HighStock({ text }: Props) {
+  const { handles } = useContextCssHandles()
+
+  return (
+    <span className={`${handles.highStockText} c-muted-2 t-body`}>{text}</span>
+  )
+}
 
 export default HighStock